+++ db.json
@@ -7,7 +7,7 @@
       "id_kelas": 1,
       "alamat": "Banyuwangi",
       "id_card": 3300133001,
-      "saldo": 1212,
+      "saldo": 5678,
       "no_tabungan": "ST-02-110-01",
       "id": "5e9c"
     },
@@ -18,7 +18,7 @@
       "id_kelas": 2,
       "alamat": "Cirebon",
       "id_card": 3300233002,
-      "saldo": 0,
+      "saldo": 1234,
       "no_tabungan": "ST-02-110-02",
       "id": "e792"
     },
